@@ -9,17 +9,17 @@ package org.openstreetmap.josm.plugins.streetside;
  */
 public interface StreetsideDataListener {
 
-  /**
-   * Fired when any image is added to the database.
-   */
-  void imagesAdded();
+    /**
+     * Fired when any image is added to the database.
+     */
+    void imagesAdded();
 
-  /**
-   * Fired when the selected image is changed by something different from
-   * manually clicking on the icon.
-   *
-   * @param oldImage Old selected {@link StreetsideAbstractImage}
-   * @param newImage New selected {@link StreetsideAbstractImage}
-   */
-  void selectedImageChanged(StreetsideAbstractImage oldImage, StreetsideAbstractImage newImage);
+    /**
+     * Fired when the selected image is changed by something different from
+     * manually clicking on the icon.
+     *
+     * @param oldImage Old selected {@link StreetsideImage}
+     * @param newImage New selected {@link StreetsideImage}
+     */
+    void selectedImageChanged(StreetsideImage oldImage, StreetsideImage newImage);
 }
