@@ -2,6 +2,7 @@
 package org.openstreetmap.josm.plugins.streetside.actions;
 
 import java.awt.event.ActionEvent;
+import java.io.Serial;
 
 import javax.swing.AbstractAction;
 
@@ -12,16 +13,17 @@ import org.openstreetmap.josm.tools.ImageProvider.ImageSizes;
 
 public class ImageReloadAction extends AbstractAction {
 
-  private static final long serialVersionUID = 7987479726049238315L;
+    @Serial
+    private static final long serialVersionUID = 7987479726049238315L;
 
-  public ImageReloadAction(final String name) {
-    super(name, ImageProvider.get("reload", ImageSizes.SMALLICON));
-  }
-
-  @Override
-  public void actionPerformed(ActionEvent arg0) {
-    if (StreetsideMainDialog.getInstance().getImage() != null) {
-      CubemapBuilder.getInstance().reload(CubemapBuilder.getInstance().getCubemap().getId());
+    public ImageReloadAction(final String name) {
+        super(name, ImageProvider.get("reload", ImageSizes.SMALLICON));
     }
-  }
+
+    @Override
+    public void actionPerformed(ActionEvent arg0) {
+        if (StreetsideMainDialog.getInstance().getImage() != null) {
+            CubemapBuilder.getInstance().reload(CubemapBuilder.getInstance().getCubemap());
+        }
+    }
 }
