@@ -10,10 +10,10 @@ package org.openstreetmap.josm.plugins.streetside.actions;
  */
 public interface WalkListener {
 
-  /**
-   * Called when a new walk thread is started.
-   *
-   * @param thread The thread executing the walk.
-   */
-  void walkStarted(WalkThread thread);
+    /**
+     * Called when a new walk thread is started.
+     *
+     * @param thread The thread executing the walk.
+     */
+    void walkStarted(WalkThread thread);
 }
