@@ -33,84 +33,78 @@ import org.openstreetmap.josm.tools.Logging;
  */
 public class StreetsideExportWriterThread extends Thread {
 
-  private static final Logger LOGGER = Logger.getLogger(StreetsideExportWriterThread.class.getCanonicalName());
+    private static final Logger LOGGER = Logger.getLogger(StreetsideExportWriterThread.class.getCanonicalName());
 
-  private final String path;
-  private final ArrayBlockingQueue<BufferedImage> queue;
-  private final ArrayBlockingQueue<StreetsideAbstractImage> queueImages;
-  private final int amount;
-  private final ProgressMonitor monitor;
+    private final ArrayBlockingQueue<BufferedImage> queue;
+    private final ArrayBlockingQueue<StreetsideAbstractImage> queueImages;
+    private final int amount;
+    private final ProgressMonitor monitor;
 
-  /**
-   * Main constructor.
-   *
-   * @param path    Path to write the pictures.
-   * @param queue     Queue of {@link StreetsideAbstractImage} objects.
-   * @param queueImages Queue of {@link BufferedImage} objects.
-   * @param amount    Amount of images that are going to be exported.
-   * @param monitor   Progress monitor.
-   */
-  public StreetsideExportWriterThread(String path, ArrayBlockingQueue<BufferedImage> queue,
-      ArrayBlockingQueue<StreetsideAbstractImage> queueImages, int amount, ProgressMonitor monitor) {
-    this.path = path;
-    this.queue = queue;
-    this.queueImages = queueImages;
-    this.amount = amount;
-    this.monitor = monitor;
-  }
-
-  @Override
-  public void run() {
-    monitor.setCustomText("Downloaded 0/" + amount);
-    BufferedImage img;
-    StreetsideAbstractImage mimg;
-    String finalPath = "";
-    for (int i = 0; i < amount; i++) {
-      try {
-        img = queue.take();
-        mimg = queueImages.take();
-
-        // Transforms the image into a byte array.
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-        ImageIO.write(img, "jpg", outputStream);
-        byte[] imageBytes = outputStream.toByteArray();
-
-        // Write EXIF tags
-        TiffOutputSet outputSet = null;
-        TiffOutputDirectory exifDirectory;
-        TiffOutputDirectory gpsDirectory;
-        // If the image is imported, loads the rest of the EXIF data.
-
-        if (null == outputSet) {
-          outputSet = new TiffOutputSet();
-        }
-        exifDirectory = outputSet.getOrCreateExifDirectory();
-        gpsDirectory = outputSet.getOrCreateGPSDirectory();
-
-        gpsDirectory.removeField(GpsTagConstants.GPS_TAG_GPS_IMG_DIRECTION_REF);
-        gpsDirectory.add(GpsTagConstants.GPS_TAG_GPS_IMG_DIRECTION_REF,
-            GpsTagConstants.GPS_TAG_GPS_IMG_DIRECTION_REF_VALUE_TRUE_NORTH);
-
-        gpsDirectory.removeField(GpsTagConstants.GPS_TAG_GPS_IMG_DIRECTION);
-        gpsDirectory.add(GpsTagConstants.GPS_TAG_GPS_IMG_DIRECTION, RationalNumber.valueOf(mimg.getMovingHe()));
-
-        exifDirectory.removeField(ExifTagConstants.EXIF_TAG_DATE_TIME_ORIGINAL);
-
-        outputSet.setGPSInDegrees(mimg.getMovingLatLon().lon(), mimg.getMovingLatLon().lat());
-        OutputStream os = new BufferedOutputStream(new FileOutputStream(finalPath + ".jpg"));
-        new ExifRewriter().updateExifMetadataLossless(imageBytes, os, outputSet);
-
-        os.close();
-      } catch (InterruptedException e) {
-        LOGGER.info("Streetside export cancelled");
-        return;
-      } catch (IOException | ImageReadException | ImageWriteException e) {
-        LOGGER.log(Logging.LEVEL_ERROR, e.getMessage(), e);
-      }
-
-      // Increases the progress bar.
-      monitor.worked(PleaseWaitProgressMonitor.PROGRESS_BAR_MAX / amount);
-      monitor.setCustomText("Downloaded " + (i + 1) + "/" + amount);
+    /**
+     * Main constructor.
+     *
+     * @param ignored    Path to write the pictures.
+     * @param queue     Queue of {@link StreetsideAbstractImage} objects.
+     * @param queueImages Queue of {@link BufferedImage} objects.
+     * @param amount    Amount of images that are going to be exported.
+     * @param monitor   Progress monitor.
+     */
+    public StreetsideExportWriterThread(String ignored, ArrayBlockingQueue<BufferedImage> queue,
+            ArrayBlockingQueue<StreetsideAbstractImage> queueImages, int amount, ProgressMonitor monitor) {
+        this.queue = queue;
+        this.queueImages = queueImages;
+        this.amount = amount;
+        this.monitor = monitor;
     }
-  }
+
+    @Override
+    public void run() {
+        monitor.setCustomText("Downloaded 0/" + amount);
+        BufferedImage img;
+        StreetsideAbstractImage mimg;
+        var finalPath = "";
+        for (var i = 0; i < amount; i++) {
+            try {
+                img = queue.take();
+                mimg = queueImages.take();
+
+                // Transforms the image into a byte array.
+                final var outputStream = new ByteArrayOutputStream();
+                ImageIO.write(img, "jpg", outputStream);
+                byte[] imageBytes = outputStream.toByteArray();
+
+                // Write EXIF tags
+                final var outputSet = new TiffOutputSet();
+                TiffOutputDirectory exifDirectory;
+                TiffOutputDirectory gpsDirectory;
+
+                exifDirectory = outputSet.getOrCreateExifDirectory();
+                gpsDirectory = outputSet.getOrCreateGPSDirectory();
+
+                gpsDirectory.removeField(GpsTagConstants.GPS_TAG_GPS_IMG_DIRECTION_REF);
+                gpsDirectory.add(GpsTagConstants.GPS_TAG_GPS_IMG_DIRECTION_REF,
+                        GpsTagConstants.GPS_TAG_GPS_IMG_DIRECTION_REF_VALUE_TRUE_NORTH);
+
+                gpsDirectory.removeField(GpsTagConstants.GPS_TAG_GPS_IMG_DIRECTION);
+                gpsDirectory.add(GpsTagConstants.GPS_TAG_GPS_IMG_DIRECTION, RationalNumber.valueOf(mimg.heading()));
+
+                exifDirectory.removeField(ExifTagConstants.EXIF_TAG_DATE_TIME_ORIGINAL);
+
+                outputSet.setGPSInDegrees(mimg.lon(), mimg.lat());
+                try (OutputStream os = new BufferedOutputStream(new FileOutputStream(finalPath + ".jpg"))) {
+                    new ExifRewriter().updateExifMetadataLossless(imageBytes, os, outputSet);
+                }
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+                LOGGER.info("Streetside export cancelled");
+                return;
+            } catch (IOException | ImageReadException | ImageWriteException e) {
+                LOGGER.log(Logging.LEVEL_ERROR, e.getMessage(), e);
+            }
+
+            // Increases the progress bar.
+            monitor.worked(PleaseWaitProgressMonitor.PROGRESS_BAR_MAX / amount);
+            monitor.setCustomText("Downloaded " + (i + 1) + "/" + amount);
+        }
+    }
 }
