@@ -2,7 +2,7 @@
 /**
  * The different modes that the {@link org.openstreetmap.josm.plugins.streetside.StreetsideLayer} can be in.
  * <br>
- * Currently there are two of them:
+ * Currently, there are two of them:
  * <ul>
  *  <li><strong>{@link org.openstreetmap.josm.plugins.streetside.mode.SelectMode SelectMode}</strong> for selecting pictures in the layer</li>
  * </ul>
