@@ -3,10 +3,8 @@ package org.openstreetmap.josm.plugins.streetside.cache;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
 
 import org.junit.jupiter.api.Test;
-import org.openstreetmap.josm.plugins.streetside.cache.StreetsideCache.Type;
 import org.openstreetmap.josm.testutils.annotations.BasicPreferences;
 
 @BasicPreferences
@@ -14,18 +12,14 @@ class StreetsideCacheTest {
 
     @Test
     void testCache() {
-        StreetsideCache cache = new StreetsideCache("00000", Type.FULL_IMAGE);
+        StreetsideCache cache = new StreetsideCache("https://ecn.t0.tiles.virtualearth.net/tiles/hs101320223333223201");
         assertNotNull(cache.getUrl());
         assertNotNull(cache.getCacheKey());
 
         assertFalse(cache.isObjectLoadable());
 
-        cache = new StreetsideCache("00000", Type.THUMBNAIL);
+        cache = new StreetsideCache("https://ecn.t0.tiles.virtualearth.net/tiles/hs101320223333223201");
         assertNotNull(cache.getCacheKey());
         assertNotNull(cache.getUrl());
-
-        cache = new StreetsideCache(null, null);
-        assertNull(cache.getCacheKey());
-        assertNull(cache.getUrl());
     }
 }
